+++ lib/src/enum/status_enum.dart
@@ -1,6 +1,8 @@
 import 'package:ansicolor/ansicolor.dart';
 
+/// Enum of Status
 enum LogStatus {
+  none,
   success,
   error,
   warning,
@@ -8,10 +10,12 @@ enum LogStatus {
   debug,
 }
 
+/// Static class helper for the enum
 class LogStatusHelper {
+  /// get the AnsiPen With the right color from the LogStatus
   static AnsiPen getColorPen(LogStatus status) {
-    var color = 16;
-    switch(status) {
+    var color = 15;
+    switch (status) {
       case LogStatus.success:
         color = 10;
         break;
@@ -27,16 +31,19 @@ class LogStatusHelper {
       case LogStatus.debug:
         color = 4;
         break;
+      case LogStatus.none:
+        color = 15;
+        break;
     }
 
     return AnsiPen()
-        ..reset()
-        ..xterm(color)
-      ;
+      ..reset()
+      ..xterm(color);
   }
 
+  /// Get the emoji from the LogStatus
   static String getEmoji(LogStatus status) {
-    switch(status) {
+    switch (status) {
       case LogStatus.success:
         return '🟢';
       case LogStatus.error:
@@ -47,16 +54,20 @@ class LogStatusHelper {
         return '🟣';
       case LogStatus.debug:
         return '🔵';
-      default :
-        return '❓';
+      case LogStatus.none:
+      default:
+        return '⚪️';
     }
   }
 
-  static String getStatusString(LogStatus status, {upperFirst = false, upperCase = false, fixedSize = true}) {
+  /// Apply some transformation on the string for better look
+  static String getStatusString(LogStatus status,
+      {upperFirst = false, upperCase = false, fixedSize = true}) {
     var statusString = _getStatusString(status, fixedSize: fixedSize);
 
     if (upperFirst) {
-      statusString = '${statusString[0].toUpperCase()}${statusString.substring(1)}';
+      statusString =
+          '${statusString[0].toUpperCase()}${statusString.substring(1)}';
     }
 
     if (upperCase) {
@@ -66,21 +77,55 @@ class LogStatusHelper {
     return statusString;
   }
 
-  static String _getStatusString(LogStatus status, { fixedSize = false }) {
-    switch(status) {
+  /// Return string terms with space for better look at output
+  static String _getStatusString(LogStatus status, {fixedSize = false}) {
+    var stringValue = '???';
+    switch (status) {
       case LogStatus.success:
-        return 'success';
-      case LogStatus.error:
-        return fixedSize ? ' error ' : 'error';
-      case LogStatus.warning:
-        return 'warning';
-      case LogStatus.info:
-        return fixedSize ? ' info  ' : 'info';
-      case LogStatus.debug:
-        return fixedSize ? ' debug ' : 'debug';
+        stringValue = 'success';
         break;
-      default :
-        return fixedSize ? '  ???  ' : '???';
+      case LogStatus.error:
+        stringValue = 'error';
+        break;
+      case LogStatus.warning:
+        stringValue = 'warning';
+        break;
+      case LogStatus.info:
+        stringValue = 'info';
+        break;
+      case LogStatus.debug:
+        stringValue = 'debug';
+        break;
+      case LogStatus.none:
+        stringValue = 'default';
+        break;
+      default:
+        stringValue = '???';
+        break;
     }
+
+    return fixedSize ? _centerPad(stringValue) : stringValue;
+  }
+
+  /// Automatic center pad string
+  static String _centerPad(String status, {wantedLength = 7}) {
+    // if status length is bigger thant the wanted size, just return.
+    if (status.length >= wantedLength) {
+      return status;
+    }
+
+    var currentLength = status.length;
+    int leftPad = ((wantedLength - currentLength) / 2).ceil();
+    int rightPad = (wantedLength - currentLength - leftPad).ceil();
+
+    for (var i = 0; i < leftPad; i++) {
+      status = ' $status';
+    }
+
+    for (var i = 0; i < rightPad; i++) {
+      status = '$status ';
+    }
+
+    return status;
   }
 }