+++ test/color_logger_test.dart
@@ -4,8 +4,14 @@ import 'package:test/test.dart';
 void main() {
   group('FixedLength', () {
     var size = 7;
+    test('None/Default', () {
+      var s = LogStatusHelper.getStatusString(LogStatus.none, fixedSize: true);
+      expect(s.length, size);
+    });
+
     test('Success', () {
-      var s = LogStatusHelper.getStatusString(LogStatus.success, fixedSize: true);
+      var s =
+          LogStatusHelper.getStatusString(LogStatus.success, fixedSize: true);
       expect(s.length, size);
     });
 
@@ -15,17 +21,20 @@ void main() {
     });
 
     test('Warning', () {
-      var s = LogStatusHelper.getStatusString(LogStatus.success, fixedSize: true);
+      var s =
+          LogStatusHelper.getStatusString(LogStatus.success, fixedSize: true);
       expect(s.length, size);
     });
 
     test('Info', () {
-      var s = LogStatusHelper.getStatusString(LogStatus.success, fixedSize: true);
+      var s =
+          LogStatusHelper.getStatusString(LogStatus.success, fixedSize: true);
       expect(s.length, size);
     });
 
     test('Debug', () {
-      var s = LogStatusHelper.getStatusString(LogStatus.success, fixedSize: true);
+      var s =
+          LogStatusHelper.getStatusString(LogStatus.success, fixedSize: true);
       expect(s.length, size);
     });
   });