+++ CHANGELOG.md
@@ -5,3 +5,9 @@
 ### 1.0.1
 
 - Readme Edit
+
+### 1.0.2
+
+- Add 'none' value for white output
+- Set 'none' as default value
+- Add comment everywhere for more pub points